backend/Grocket/api/users/views.py
from djoser import views as djviews
from rest_framework import permissions


class CustomUserRetrieveViewSet(djviews.UserViewSet):
    permission_classes = (permissions.AllowAny,)


class CustomUserRegisterViewSet(djviews.UserViewSet):
    pass
